refactor tests to mirror src
