+++ skore/tests/unit/sklearn/train_test_split/test_train_test_split.py
@@ -1,32 +1,14 @@
 import contextlib
 
-import numpy
-import pandas
 import pytest
 from skore.sklearn.train_test_split.train_test_split import (
-    HighClassImbalanceWarning,
     train_test_split,
 )
-
-target = [0] * 100 + [1] * 100 + [2] * 300
-
-
-@pytest.mark.parametrize(
-    "y",
-    [
-        pytest.param(target, id="list"),
-        pytest.param(pandas.Series(target), id="pandas-series"),
-        pytest.param(numpy.array(target), id="numpy"),
-    ],
+from skore.sklearn.train_test_split.warning import (
+    HighClassImbalanceTooFewExamplesWarning,
+    HighClassImbalanceWarning,
 )
-def test_check_high_class_imbalance(y):
-    check = HighClassImbalanceWarning.check(
-        y=y,
-        stratify=None,
-        ml_task="multiclass-classification",
-    )
 
-    assert check is False
 
 
 def test_train_test_split_warns():
